chore: fetching balances through module
